feat: getTable 함수 분리, 유저 예약 내역 파싱
